refactor(runtime-host): split execution composition
